Clean up exception utilities
