+++ grapesy/test-disconnect/Test/Disconnect/Util/Exception.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE CPP #-}
 
--- | Exception utilities
-module Test.Util.Exception (
+module Test.Disconnect.Util.Exception (
     testFormatCtx
   , uncaughtExceptionHandler
   ) where
@@ -10,19 +9,15 @@ import Control.Concurrent
 import Control.Exception
 import Data.Function ((&))
 import Data.Maybe (fromMaybe)
-import Data.Proxy
 import System.IO
-
-import Network.HTTP2.Client qualified as HTTP2
 
 #if MIN_VERSION_base(4,18,0)
 import GHC.Conc.Sync (threadLabel)
 #endif
 
-import Network.GRPC.Common.Exception
+import Network.HTTP2.Client qualified as HTTP2
 
-import Test.Driver.ClientServer (FirstTestFailure)
-import Test.Prop.Dialogue (RegressionTestFailed)
+import Network.GRPC.Common.Exception
 
 {-------------------------------------------------------------------------------
   Exception rendering
@@ -31,15 +26,13 @@ import Test.Prop.Dialogue (RegressionTestFailed)
 testFormatCtx :: FormatCtx
 testFormatCtx = grapesyFormatCtx
     & insertFormatCtx http2
-    & insertFormatCtx_ (Proxy @FirstTestFailure)
-    & insertFormatCtx_ (Proxy @RegressionTestFailed)
   where
     http2 :: FormatCtx -> HTTP2.HTTP2Error -> Doc
     http2 ctx = \case
       HTTP2.BadThingHappen se ->
         withHeader "BadThingHappen" $ toExceptionDoc ctx se
       other ->
-        fromLines (displayException other)
+        fromLines $ displayException other
 
 {-------------------------------------------------------------------------------
   Uncaught exception handler
