+++ grapesy/test-grapesy/Main.hs
@@ -1,22 +1,9 @@
-{-# LANGUAGE CPP #-}
-
 module Main (main) where
 
-import Control.Concurrent
-import Control.Exception
-import Data.Function ((&))
-import Data.Maybe (fromMaybe)
 import GHC.Conc (setUncaughtExceptionHandler)
-import System.IO
 import Test.Tasty
 
-#if MIN_VERSION_base(4,18,0)
-import GHC.Conc.Sync (threadLabel)
-#endif
-
-import Network.HTTP2.Client qualified as HTTP2
-
-import Network.GRPC.Common.Exception
+import Test.Util.Exception
 
 import Test.Common.Exception                  qualified as Exception
 import Test.Prop.Dialogue                     qualified as Dialogue
@@ -61,32 +48,3 @@ main = do
             Exception.tests
           ]
       ]
-
-uncaughtExceptionHandler :: SomeException -> IO ()
-uncaughtExceptionHandler e = do
-    tid    <- myThreadId
-    mLabel :: Maybe String <-
-#if MIN_VERSION_base(4,18,0)
-      threadLabel tid
-#else
-      return $ Just "unknown label"
-#endif
-    hPutStrLn stderr $ concat [
-         "Uncaught exception in "
-      , show tid
-      , " ("
-      , fromMaybe "unlabelled" mLabel
-      , "): "
-      , renderException formatCtx e
-      ]
-  where
-    formatCtx :: FormatCtx
-    formatCtx = defaultFormatCtx
-        & insertFormatCtx http2
-
-    http2 :: FormatCtx -> HTTP2.HTTP2Error -> Doc
-    http2 ctx = \case
-      HTTP2.BadThingHappen se ->
-        withHeader "BadThingHappen" $ toExceptionDoc ctx se
-      other ->
-        fromLines (displayException other)
